+++ application/auth/src/main/java/animal/auth/domain/CompanyUser.java
@@ -1,19 +1,15 @@
-package animal.domain;
+package animal.auth.domain;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
-@Table(name = "p_delivery")
+@Table(name = "p_company")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class DeliveryUser extends User {
+public class CompanyUser extends User {
 
   @Column(nullable = false)
-  private DeliveryType type;
-
-  @Column(nullable = false)
-  private String slackId;
+  private CompanyType type;
 
 }
