+++ application/auth/src/main/java/animal/auth/domain/DeliveryType.java
@@ -1,4 +1,4 @@
-package animal.domain;
+package animal.auth.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
